Moved create comments method to server side.
