+++ server/documents/comment.js
@@ -1,7 +1,13 @@
 import {check, Match} from 'meteor/check';
 import {Meteor} from 'meteor/meteor';
 
+import {Node} from 'prosemirror-model';
+
 import {Comment} from '/lib/documents/comment';
+import {schema} from '/lib/simple-schema.js';
+import {User} from './user';
+import {Document} from './document';
+import {Activity} from "./activity";
 
 Meteor.publish('Comment.list', function commentList(args) {
   check(args, {
@@ -129,6 +135,92 @@ Meteor.methods({
       multi: true,
     });
   },
+
+  'Comment.create'(args) {
+    check(args, {
+      documentId: Match.DocumentId,
+      highlightKey: Match.DocumentId,
+      body: Object,
+      replyTo: Match.Maybe(Match.DocumentId),
+    });
+
+    // Validate body.
+    Node.fromJSON(schema, args.body).check();
+
+    const user = Meteor.user(User.REFERENCE_FIELDS());
+
+    // We need user reference.
+    if (!user || !user.hasPermission(Comment.PERMISSIONS.CREATE)) {
+      throw new Meteor.Error('unauthorized', "Unauthorized.");
+    }
+
+    const document = Document.documents.findOne(Document.restrictQuery({
+      _id: args.documentId,
+    }, Document.PERMISSIONS.COMMENT_CREATE, user), {fields: Document.REFERENCE_FIELDS()});
+    if (!document) {
+      throw new Meteor.Error('not-found', `Document cannot be found.`);
+    }
+
+    let replyTo = null;
+    if (args.replyTo) {
+      replyTo = Comment.documents.findOne(Comment.restrictQuery({
+        _id: args.replyTo,
+      }, Comment.PERMISSIONS.SEE, user), {fields: Document.REFERENCE_FIELDS()});
+      if (!replyTo) {
+        throw new Meteor.Error('not-found', `Comment cannot be found.`);
+      }
+    }
+
+    const createdAt = new Date();
+
+    const commentId = Comment.documents.insert({
+      createdAt,
+      author: user.getReference(),
+      document: document.getReference(),
+      body: args.body,
+      versionFrom: null,
+      versionTo: null,
+      // TODO: Validate highlight key.
+      highlightKey: args.highlightKey,
+      replyTo: replyTo && replyTo.getReference(),
+      status: Comment.STATUS.CREATED,
+    });
+
+    Document.documents.update({
+      _id: args.documentId,
+      lastActivity: {
+        $lt: createdAt,
+      },
+    }, {
+      $set: {
+        lastActivity: createdAt,
+      },
+    });
+
+    if (Meteor.isServer) {
+      Activity.documents.insert({
+        timestamp: createdAt,
+        connection: this.connection.id,
+        byUser: user.getReference(),
+        // We inform all followers of this document.
+        // TODO: Implement once we have followers.
+        forUsers: [],
+        type: 'commentCreated',
+        level: Activity.LEVEL.GENERAL,
+        data: {
+          document: document.getReference(),
+          comment: {
+            _id: commentId,
+          },
+        },
+      });
+    }
+
+    return {
+      _id: commentId,
+    };
+  },
+
 });
 
 // For testing.
